CRM Double Tap Sign-In-Sign Up
In this project, you will be tasked with creating a new website for a “Fake CRM”. This website’s mock design is also included in this file. This will be a group project in which each person should share an equal role in the finished product.

One group member (Original Team Member) will need to create the starter files in their Github Profile. From there each person will need to fork & clone the repository and begin working on their portion of the product. When each of you has completed your parts of the project, commit them back to github and submit a pull request to the original team member’s repository. Each person will then need to submit the original team member’s github link to google classroom.